+++ blink/address.h
@@ -4,11 +4,11 @@
 #include "blink/machine.h"
 #include "blink/modrm.h"
 
-u64 AddressOb(struct Machine *, u32);
-u64 AddressDi(struct Machine *, u32);
-u64 AddressSi(struct Machine *, u32);
-u64 *GetSegment(struct Machine *, u32, int);
-u64 DataSegment(struct Machine *, u32, u64);
+u64 AddressOb(struct Machine *, u64);
+u64 AddressDi(struct Machine *, u64);
+u64 AddressSi(struct Machine *, u64);
+u64 *GetSegment(struct Machine *, u64, int);
+u64 DataSegment(struct Machine *, u64, u64);
 
 static inline u64 MaskAddress(u32 mode, u64 x) {
   if (mode != XED_MODE_LONG) {
@@ -21,7 +21,7 @@ static inline u64 MaskAddress(u32 mode, u64 x) {
   return x;
 }
 
-static inline u64 AddSegment(struct Machine *m, u32 rde, u64 i, u64 s) {
+static inline u64 AddSegment(struct Machine *m, u64 rde, u64 i, u64 s) {
   if (!Sego(rde)) {
     return i + s;
   } else {
