+++ blink/rdrand.c
@@ -41,12 +41,12 @@ static u64 Vigna(u64 s[1]) {
   return z ^ (z >> 31);
 }
 
-static void OpRand(struct Machine *m, u32 rde, u64 x) {
+static void OpRand(struct Machine *m, u64 rde, u64 x) {
   WriteRegister(rde, RegRexbRm(m, rde), x);
   m->flags = SetFlag(m->flags, FLAGS_CF, true);
 }
 
-void OpRdrand(struct Machine *m, u32 rde) {
+void OpRdrand(struct Machine *m, u64 rde) {
   pthread_mutex_lock(&g_rdrand.lock);
   if (!(g_rdrand.count++ % RESEED_INTERVAL)) {
     unassert(GetRandom(&g_rdrand.state, 8) == 8);
@@ -55,7 +55,7 @@ void OpRdrand(struct Machine *m, u32 rde) {
   pthread_mutex_unlock(&g_rdrand.lock);
 }
 
-void OpRdseed(struct Machine *m, u32 rde) {
+void OpRdseed(struct Machine *m, u64 rde) {
   u64 x;
   unassert(GetRandom(&x, 8) == 8);
   OpRand(m, rde, x);
