+++ blink/stack.c
@@ -52,7 +52,7 @@ static const u8 kCallOsz[2][3] = {
     },
 };
 
-static void WriteStackWord(u8 *p, u32 rde, u32 osz, u64 x) {
+static void WriteStackWord(u8 *p, u64 rde, u32 osz, u64 x) {
   IGNORE_RACES_START();
   if (osz == 8) {
     Write64(p, x);
@@ -78,7 +78,7 @@ static u64 ReadStackWord(u8 *p, u32 osz) {
   return x;
 }
 
-static void PushN(struct Machine *m, u32 rde, u64 x, unsigned mode,
+static void PushN(struct Machine *m, u64 rde, u64 x, unsigned mode,
                   unsigned osz) {
   u8 *w;
   u64 v;
@@ -109,16 +109,16 @@ static void PushN(struct Machine *m, u32 rde, u64 x, unsigned mode,
   IGNORE_RACES_END();
 }
 
-void Push(struct Machine *m, u32 rde, u64 x) {
+void Push(struct Machine *m, u64 rde, u64 x) {
   PushN(m, rde, x, Eamode(rde), kStackOsz[Osz(rde)][Mode(rde)]);
 }
 
-void OpPushZvq(struct Machine *m, u32 rde) {
+void OpPushZvq(struct Machine *m, u64 rde) {
   unsigned osz = kStackOsz[Osz(rde)][Mode(rde)];
   PushN(m, rde, ReadStackWord(RegRexbSrm(m, rde), osz), Eamode(rde), osz);
 }
 
-static u64 PopN(struct Machine *m, u32 rde, u16 extra, unsigned osz) {
+static u64 PopN(struct Machine *m, u64 rde, u16 extra, unsigned osz) {
   u64 v;
   void *p[2];
   u8 b[8];
@@ -143,11 +143,11 @@ static u64 PopN(struct Machine *m, u32 rde, u16 extra, unsigned osz) {
   return ReadStackWord(AccessRam(m, v, osz, p, b, true), osz);
 }
 
-u64 Pop(struct Machine *m, u32 rde, u16 extra) {
+u64 Pop(struct Machine *m, u64 rde, u16 extra) {
   return PopN(m, rde, extra, kStackOsz[Osz(rde)][Mode(rde)]);
 }
 
-void OpPopZvq(struct Machine *m, u32 rde) {
+void OpPopZvq(struct Machine *m, u64 rde) {
   u64 x;
   unsigned osz;
   osz = kStackOsz[Osz(rde)][Mode(rde)];
@@ -165,30 +165,30 @@ void OpPopZvq(struct Machine *m, u32 rde) {
   }
 }
 
-static void OpCall(struct Machine *m, u32 rde, u64 func) {
+static void OpCall(struct Machine *m, u64 rde, u64 func) {
   PushN(m, rde, m->ip, Mode(rde), kCallOsz[Osz(rde)][Mode(rde)]);
   m->ip = func;
 }
 
-void OpCallJvds(struct Machine *m, u32 rde) {
+void OpCallJvds(struct Machine *m, u64 rde) {
   OpCall(m, rde, m->ip + m->xedd->op.disp);
 }
 
-static u64 LoadAddressFromMemory(struct Machine *m, u32 rde) {
+static u64 LoadAddressFromMemory(struct Machine *m, u64 rde) {
   unsigned osz;
   osz = kCallOsz[Osz(rde)][Mode(rde)];
   return ReadStackWord(GetModrmRegisterWordPointerRead(m, rde, osz), osz);
 }
 
-void OpCallEq(struct Machine *m, u32 rde) {
+void OpCallEq(struct Machine *m, u64 rde) {
   OpCall(m, rde, LoadAddressFromMemory(m, rde));
 }
 
-void OpJmpEq(struct Machine *m, u32 rde) {
+void OpJmpEq(struct Machine *m, u64 rde) {
   m->ip = LoadAddressFromMemory(m, rde);
 }
 
-void OpLeave(struct Machine *m, u32 rde) {
+void OpLeave(struct Machine *m, u64 rde) {
   switch (Eamode(rde)) {
     case XED_MODE_LONG:
       Put64(m->sp, Get64(m->bp));
@@ -207,25 +207,25 @@ void OpLeave(struct Machine *m, u32 rde) {
   }
 }
 
-void OpRet(struct Machine *m, u32 rde) {
+void OpRet(struct Machine *m, u64 rde) {
   m->ip = Pop(m, rde, m->xedd->op.uimm0);
 }
 
-void OpPushEvq(struct Machine *m, u32 rde) {
+void OpPushEvq(struct Machine *m, u64 rde) {
   unsigned osz;
   osz = kStackOsz[Osz(rde)][Mode(rde)];
   Push(m, rde,
        ReadStackWord(GetModrmRegisterWordPointerRead(m, rde, osz), osz));
 }
 
-void OpPopEvq(struct Machine *m, u32 rde) {
+void OpPopEvq(struct Machine *m, u64 rde) {
   unsigned osz;
   osz = kStackOsz[Osz(rde)][Mode(rde)];
   WriteStackWord(GetModrmRegisterWordPointerWrite(m, rde, osz), rde, osz,
                  Pop(m, rde, 0));
 }
 
-static void Pushaw(struct Machine *m, u32 rde) {
+static void Pushaw(struct Machine *m, u64 rde) {
   u16 v;
   u8 b[8][2];
   memcpy(b[0], m->di, 2);
@@ -240,7 +240,7 @@ static void Pushaw(struct Machine *m, u32 rde) {
   VirtualRecv(m, m->ss + v, b, sizeof(b));
 }
 
-static void Pushad(struct Machine *m, u32 rde) {
+static void Pushad(struct Machine *m, u64 rde) {
   u32 v;
   u8 b[8][4];
   memcpy(b[0], m->di, 4);
@@ -255,7 +255,7 @@ static void Pushad(struct Machine *m, u32 rde) {
   VirtualRecv(m, m->ss + v, b, sizeof(b));
 }
 
-static void Popaw(struct Machine *m, u32 rde) {
+static void Popaw(struct Machine *m, u64 rde) {
   u8 b[8][2];
   VirtualSend(m, b, m->ss + Read16(m->sp), sizeof(b));
   Put16(m->sp, (Get32(m->sp) + sizeof(b)) & 0xffff);
@@ -269,7 +269,7 @@ static void Popaw(struct Machine *m, u32 rde) {
   memcpy(m->ax, b[7], 2);
 }
 
-static void Popad(struct Machine *m, u32 rde) {
+static void Popad(struct Machine *m, u64 rde) {
   u8 b[8][4];
   VirtualSend(m, b, m->ss + Get32(m->sp), sizeof(b));
   Put64(m->sp, (Get32(m->sp) + sizeof(b)) & 0xffffffff);
@@ -283,7 +283,7 @@ static void Popad(struct Machine *m, u32 rde) {
   memcpy(m->ax, b[7], 4);
 }
 
-void OpPusha(struct Machine *m, u32 rde) {
+void OpPusha(struct Machine *m, u64 rde) {
   switch (Eamode(rde)) {
     case XED_MODE_REAL:
       Pushaw(m, rde);
@@ -298,7 +298,7 @@ void OpPusha(struct Machine *m, u32 rde) {
   }
 }
 
-void OpPopa(struct Machine *m, u32 rde) {
+void OpPopa(struct Machine *m, u64 rde) {
   switch (Eamode(rde)) {
     case XED_MODE_REAL:
       Popaw(m, rde);
@@ -313,7 +313,7 @@ void OpPopa(struct Machine *m, u32 rde) {
   }
 }
 
-void OpCallf(struct Machine *m, u32 rde) {
+void OpCallf(struct Machine *m, u64 rde) {
   Push(m, rde, m->cs >> 4);
   Push(m, rde, m->ip);
   m->cs = m->xedd->op.uimm0 << 4;
@@ -323,7 +323,7 @@ void OpCallf(struct Machine *m, u32 rde) {
   }
 }
 
-void OpRetf(struct Machine *m, u32 rde) {
+void OpRetf(struct Machine *m, u64 rde) {
   m->ip = Pop(m, rde, 0);
   m->cs = Pop(m, rde, m->xedd->op.uimm0) << 4;
   if (m->system->onlongbranch) {
